+++ src/containers/EntityTable/config.js
@@ -7,20 +7,10 @@ import { statusColumn } from './columns/status';
 import { listTypeColumn } from './columns/listType';
 import { descriptionColumn } from './columns/description';
 import { metricTypeColumn } from './columns/metricType';
-import { metricNameColumn } from './columns/metricName';
-import { camelCaseToRegularForm } from '../../utils/string';
 
-export const getHelpLink = entityName => {
-  switch (entityName) {
-    case 'lists':
-      return 'docs.cxengage.net/Help/Content/Configuration/Lists/Lists.htm';
-    case 'emailTemplates':
-      return 'docs.cxengage.net/Help/Content/Configuration/Email_Templates/Updating_Email_Templates.htm';
-    default:
-      return undefined;
-  }
-};
-
+/**
+ * Returns table columns for current entity
+ */
 export function getTableColumns(entityName) {
   switch (entityName) {
     case 'lists':
@@ -32,25 +22,11 @@ export function getTableColumns(entityName) {
     case 'outboundIdentifierLists':
       return [nameColumn, descriptionColumn];
     case 'customMetrics':
-      return [
-        metricTypeColumn,
-        metricNameColumn,
-        descriptionColumn,
-        statusColumn
-      ];
+      return [metricTypeColumn, nameColumn, descriptionColumn, statusColumn];
     case 'chatWidgets':
       return [nameColumn, descriptionColumn, statusColumn];
     //hygen-inject-before
     default:
       return [];
-  }
-}
-
-export function getTitle(entityName) {
-  switch (entityName) {
-    case 'customMetrics':
-      return 'Statistics Management';
-    default:
-      return camelCaseToRegularForm(entityName);
   }
 }