+++ src/containers/SidePanelHeader/index.js
@@ -11,7 +11,7 @@ import * as MODALS from '../ConfirmationDialog/constants.js';
 
 import { capitalizeFirstLetter } from '../../utils/string';
 
-import { unsetSelectedEntityId, setConfirmationDialog } from '../../redux/modules/entities';
+import { unsetSelectedEntityId, setConfirmationDialog, copyCurrentEntity } from '../../redux/modules/entities';
 import {
   getCurrentEntity,
   getSelectedEntityId,
@@ -21,14 +21,16 @@ import {
   getSelectedEntityBulkChangeItems
 } from '../../redux/modules/entities/selectors';
 
-export function mapDispatchToProps(dispatch) {
+export function mapDispatchToProps(dispatch, ownProps) {
+  let actions = {};
+  // TODO: copy button is experimental feature and is feature flagged.
+  if (location.hash.includes('alpha')) {
+    actions.copy = () => dispatch(copyCurrentEntity());
+  }
   return {
-    onClose: () => {
-      dispatch(unsetSelectedEntityId());
-    },
-    onToggle: () => {
-      dispatch(setConfirmationDialog(MODALS.CONFIRM_ENTITY_ACTIVE_TOGGLE));
-    }
+    ...actions,
+    onClose: () => dispatch(unsetSelectedEntityId()),
+    onToggle: () => dispatch(setConfirmationDialog(MODALS.CONFIRM_ENTITY_ACTIVE_TOGGLE))
   };
 }