+++ src/redux/modules/entities/permissions/selectors.js
@@ -1,0 +1,5 @@
+/*
+ * Copyright © 2015-2018 Serenova, LLC. All rights reserved.
+ */
+
+export const getAllPermissions = state => state.getIn(['Entities', 'permissions', 'data']);